+++ packages/app-runtime/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fayz-ai/app-runtime
 
+## 0.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @fayz-ai/ui@0.6.0
+  - @fayz-ai/core@0.6.0
+  - @fayz-ai/saas@0.6.0
+  - @fayz-ai/storefront@0.6.0
+  - @fayz-ai/shop@0.6.0
+  - @fayz-ai/auth@0.6.0
+
 ## 0.5.0
 
 ### Minor Changes