+++ plugins/plugin-financial/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @fayz-ai/plugin-financial
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @fayz-ai/ui@0.6.0
+  - @fayz-ai/core@0.6.0
+  - @fayz-ai/saas@0.6.0
+
 ## 0.1.6
 
 ### Patch Changes