+++ plugins/plugin-orders/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @fayz-ai/plugin-orders
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @fayz-ai/sdk@0.6.0
+  - @fayz-ai/core@0.6.0
+  - @fayz-ai/saas@0.6.0
+
 ## 0.2.2
 
 ### Patch Changes