+++ plugins/plugin-shop/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @fayz-ai/plugin-shop
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @fayz-ai/ui@0.6.0
+  - @fayz-ai/core@0.6.0
+  - @fayz-ai/saas@0.6.0
+  - @fayz-ai/shop@0.6.0
+
 ## 0.2.3
 
 ### Patch Changes